Move Swap and Stake under Monetization
