+++ bitrise.yml
@@ -162,10 +162,8 @@ stages:
       - run_tag_smoke_confirmations_android: {}
       - run_tag_smoke_confirmations_redesigned_ios: {}
       - run_tag_smoke_ramps_android: {}
-      - run_tag_smoke_swaps_ios: {}
-      - run_tag_smoke_swaps_android: {}
-      - run_tag_smoke_stake_ios: {}
-      - run_tag_smoke_stake_android: {}
+      - run_tag_smoke_monetization_ios: {}
+      - run_tag_smoke_monetization_android: {}
       - run_tag_smoke_core_ios: {}
       - run_tag_smoke_core_android: {}
       - run_tag_multichain_permissions_ios: {}
@@ -192,8 +190,8 @@ stages:
       # - run_tag_smoke_identity_android: {}
       # - run_tag_smoke_assets_ios: {}
       - run_tag_smoke_assets_android: {}
-      # - run_tag_smoke_swaps_ios: {}
-      # - run_tag_smoke_swaps_android: {}
+      # - run_tag_smoke_monetization_ios: {}
+      # - run_tag_smoke_monetization_android: {}
       - run_tag_smoke_core_ios: {}
       - run_tag_smoke_core_android: {}
       - run_tag_upgrade_android: {}
@@ -660,41 +658,22 @@ workflows:
       - TEST_SUITE_TAG: '.*SmokeConfirmationsRedesigned.*'
     after_run:
       - ios_e2e_test
-  run_tag_smoke_swaps_ios:
+  run_tag_smoke_monetization_ios:
     envs:
-      - TEST_SUITE_FOLDER: './e2e/specs/swaps/*'
-      - TEST_SUITE_TAG: '.*SmokeSwaps.*'
+      - TEST_SUITE_FOLDER: './e2e/specs/monetization/*'
+      - TEST_SUITE_TAG: '.*SmokeMonetization.*'
     after_run:
       - ios_e2e_test
-  run_tag_smoke_swaps_android:
+  run_tag_smoke_monetization_android:
     meta:
       bitrise.io:
         stack: linux-docker-android-22.04
         machine_type_id: elite-xl
     envs:
-      - TEST_SUITE_FOLDER: './e2e/specs/swaps/*'
-      - TEST_SUITE_TAG: '.*SmokeSwaps.*'
+      - TEST_SUITE_FOLDER: './e2e/specs/monetization/*'
+      - TEST_SUITE_TAG: '.*SmokeMonetization.*'
     after_run:
       - android_e2e_test
-  run_tag_smoke_stake_ios:
-    envs:
-      - TEST_SUITE_FOLDER: './e2e/specs/stake/*'
-      - TEST_SUITE_TAG: '.*SmokeStake.*'
-    after_run:
-      - ios_e2e_test
-  run_tag_smoke_stake_android:
-    meta:
-      bitrise.io:
-        stack: linux-docker-android-22.04
-        machine_type_id: elite-xl
-    envs:
-      - TEST_SUITE_FOLDER: './e2e/specs/stake/*'
-      - TEST_SUITE_TAG: '.*SmokeStake.*'
-    after_run:
-      - android_e2e_test
-  run_ios_api_specs:
-    after_run:
-      - ios_api_specs
   run_tag_smoke_core_ios:
     envs:
       - TEST_SUITE_FOLDER: './e2e/spec/*/**/*'
