+++ e2e/specs/monetization/swaps/swap-action-regression.spec.js
