+++ e2e/specs/monetization/swaps/swap-token-chart.spec.js
